requirements.txt
accelerate==0.29.3
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
anytree==2.12.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
entrypoints==0.4
exceptiongroup==1.2.1
Faker==24.7.1
favicon==0.7.0
filelock==3.13.4
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
isodate==0.6.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
llama_cpp_python==0.2.67
lmql==0.7.3
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==1.23.6
packaging==24.0
pandas==2.2.1
pillow==10.3.0
prometheus_client==0.20.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.2
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.8.1b0
Pygments==2.17.2
pymdown-extensions==10.7.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.34.0
regex==2024.4.28
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
st-annotated-text==4.0.1
st-cytoscape==0.0.5
st-mui-dialog==0.0.10
streamlit-agraph==0.0.45
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-cytoscapejs==0.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-float==0.3.2
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-modal==0.1.2
streamlit-nightly==1.33.1.dev20240421
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.6.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.2
transformers==4.40.1
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.0
watchdog==4.0.0
yarl==1.9.4
